Configuration via flags
This change-set overhauls the configuration of eagle. Instead of relying
on configuration files in the Toml format it receives a set of flags. As
a consequence eagle is now limited to a single `Test` which encapsulates
the configuration and interacts with a set of `Targets` which do the
actual load testing. Currently the following flags are supported:

```
-test.name="unknown": Name of the test to run.
-test.path="/": Path to hit on the targets
-test.rate=100: Number of requests to send during test duration.
-test.target=: Target to hit by the test with the following format: -test.target="NAME address/url"
```

Additionally this change collapses a lot of code and takes care of
consistent naming.